requirements.txt
PySide6
pexpect
pyusb
psutil
